Changed test setup
