+++ src/test/resources/h2test.properties
@@ -2,5 +2,3 @@ spring.H2.database-driver=org.h2.Driver
 spring.H2.datasource.url=jdbc:h2:mem:test
 spring.H2.datasource.username=sa
 spring.H2.datasource.password=sa
-
-test.reset.sql.template=ALTER TABLE %s ALTER COLUMN id RESTART WITH 1
